[feat] slider Component 수정 아직 미완성
